// idemp-rest-test/testing/src/types/org.bandahealth.idempiere.rest.d.ts
export interface BaseEntity extends BaseMetadata {
    description: string;
    name: string;
    value: string;
}

export interface BaseListResponse<T> {
    results: T[];
    pagingInfo: Paging;
}

export interface BaseMetadata extends Serializable {
    clientId: number;
    orgId: number;
    uuid: string;
    isActive: boolean;
    created: string;
    createdTimestamp: Date;
}

export interface BaseObject extends Serializable {
    id: number;
    name: string;
}

export interface BusinessPartner extends BaseEntity {
    totalOpenBalance: number;
}
// ...
export interface CodedDiagnosis extends BaseMetadata {
    cielName: string;
    cielId: number;
    conceptClass: string;
    icd10: string;
    synomedCT: number;
    synomedNP: number;
    synonyms: string;
    description: string;
    moh705bGreaterThan5: string;
    searchTerms: string;
    moh705aLessthan5: string;
    moh705aLessThan5: string;
}
// ...
export interface NHIF extends Serializable {
    type: NHIFType;
    relationship: NHIFRelationship;
    claimNumber: string;
    memberId: string;
    number: string;
    memberName: string;
}

export interface NHIFRelationship extends BaseEntity {
}

export interface NHIFType extends BaseEntity {
}

export interface Order extends BaseMetadata {
    businessPartner: BusinessPartner;
    dateOrdered: string;
    grandTotal: number;
    description: string;
    orderLines: OrderLine[];
    payments: Payment[];
    docStatus: string;
    voidedReason: VoidedReason;
    warehouse: Warehouse;
}

export interface OrderLine extends BaseMetadata {
    expenseCategory: ExpenseCategory;
    orderId: number;
    product: Product;
    price: number;
    quantity: number;
    lineNetAmount: number;
    instructions: string;
    charge: Charge;
    chargeInformationList: OrderLineChargeInformation[];
    description: string;
    attributeSetInstance: AttributeSetInstance;
}
// ...
export interface Patient extends BusinessPartner {
    patientNumber: string;
    dateOfBirth: string;
    phone: string;
    address: string;
    gender: string;
    email: string;
    nhifRelationship: string;
    nhifMemberName: string;
    nhifNumber: string;
    nhifType: string;
    nationalId: string;
    occupation: string;
    nextOfKinName: string;
    nextOfKinContact: string;
    localPatientNumber: string;
    totalVisits: number;
    lastVisitDate: string;
    approximateDateOfBirth: boolean;
    isApproximateDateOfBirth: boolean;
}

export interface PatientSummary extends BaseMetadata {
    totalVisits: number;
    inpatientVisits: number;
    outpatientVisits: number;
    patientsWithOutstandingBalance: number;
}

export interface PatientType extends BaseEntity {
}

export interface Payment extends BaseMetadata {
    patient: Patient;
    chargeId: number;
    orderId: number;
    payAmount: number;
    paymentType: PaymentType;
    description: string;
    nhif: NHIF;
    docStatus: string;
    transactionDate: string;
    tenderAmount: number;
}

export interface PaymentType extends BaseEntity {
}
// ...
export interface ProcessStage extends BaseEntity {
}
// ...
export interface Referral extends BaseEntity {
}

export interface Role extends BaseEntity {
}
// ...
export interface User extends BaseMetadata {
    name: string;
    dateLastLogin: Date;
    roles: Role[];
    resetPassword: string;
}

export interface Vendor extends BusinessPartner {
    phoneNumber: string;
    emailAddress: string;
}

export interface Visit extends Order {
    newVisit: boolean;
    clinicalNotes: string;
    labNotes: string;
    patientType: PatientType;
    referral: Referral;
    status: OrderStatus;
    patient: Patient;
    chiefComplaint: string;
    temperature: string;
    pulse: string;
    respiratoryRate: string;
    height: string;
    weight: string;
    clinician: User;
    processStage: ProcessStage;
    referredFromTo: string;
    visitDate: Date;
    primaryCodedDiagnosis: CodedDiagnosis;
    secondaryCodedDiagnosis: CodedDiagnosis;
    primaryUnCodedDiagnosis: string;
    secondaryUnCodedDiagnosis: string;
    systolicBloodPressure: number;
    diastolicBloodPressure: number;
    oxygenSaturation: number;
    diagnosis: string;
}

export interface VoidedReason extends BaseEntity {
    lineNo: number;
}

export interface Warehouse extends BaseEntity {
    defaultWarehouse: boolean;
    locators: Locator[];
}

export interface Window extends BaseEntity {
}

export interface Serializable {
}

export type OrderStatus = "WAITING" | "DISPENSING" | "PENDING" | "PENDING_COMPLETION" | "COMPLETED";